pyproject.toml
[project]
name = "ai-document-summarizer"
version = "1.0.0"
description = "AI Document Summarizer - анализ документов с помощью LLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dependency-injector>=4.48.3",
    "easyocr>=1.7.2",
    "faster-whisper>=1.2.1",
    "fpdf>=1.7.2",
    "loguru>=0.7.3",
    "opencv-python-headless>=4.13.0.92",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "tenacity>=9.1.4",
    "toml>=0.10.2",
    "typer>=0.24.0",
]

[dependency-groups]
dev = ["pytest>=9.0.2", "pytest-mock>=3.15.1", "requests-mock>=1.12.1"]

[tool.ruff]
line-length = 90
target-version = "py313"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = []
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
